Move review applications view and new applications dialog to live under activity folder
