+++ bitwarden_license/bit-web/src/app/dirt/access-intelligence/v2/activity-tab/activity-tab.component.spec.ts
@@ -19,12 +19,11 @@ import { OrganizationId } from "@bitwarden/common/types/guid";
 import { CipherView } from "@bitwarden/common/vault/models/view/cipher.view";
 import { DialogRef, DialogService } from "@bitwarden/components";
 
+import { ActivityTabComponent } from "./activity-tab.component";
 import {
   NewApplicationsDialogResultType,
   NewApplicationsDialogV2Component,
-} from "../new-applications-dialog-v2/new-applications-dialog-v2.component";
-
-import { ActivityTabComponent } from "./activity-tab.component";
+} from "./new-applications-dialog-v2/new-applications-dialog-v2.component";
 
 /**
  * Mock type for AccessIntelligenceDataService that uses BehaviorSubjects
