+++ bitwarden_license/bit-web/src/app/dirt/access-intelligence/v2/activity-tab/new-applications-dialog-v2/review-applications-view-v2/review-applications-view-v2.component.stories.ts
@@ -57,7 +57,7 @@ const mockCiphers = [
 ];
 
 export default {
-  title: "DIRT/Access Intelligence/Review Applications",
+  title: "DIRT/Access Intelligence/Activity Tab/Review Applications",
   component: ReviewApplicationsViewV2Component,
   decorators: [
     moduleMetadata({
