+++ bitwarden_license/bit-web/src/app/dirt/access-intelligence/v2/activity-tab/new-applications-dialog-v2/review-applications-view-v2/review-applications-view-v2.component.ts
