+++ bitwarden_license/bit-web/src/app/dirt/access-intelligence/v2/activity-tab/password-change-metric-v2/password-change-metric-v2.component.stories.ts
@@ -40,7 +40,7 @@ function createTask(id: string, cipherId: string, status: SecurityTaskStatus): S
 }
 
 export default {
-  title: "DIRT/Access Intelligence/Password Change Metric",
+  title: "DIRT/Access Intelligence/Activity Tab/Password Change Metric",
   component: PasswordChangeMetricV2Component,
   decorators: [
     moduleMetadata({
